requirements.txt
asgiref==3.5.2
backports.zoneinfo==0.2.1
cachetools==5.2.0
certifi==2022.6.15
charset-normalizer==2.1.0
Django==4.0.6
google-api-core==2.8.2
google-api-python-client==2.52.0
google-auth==2.9.1
google-auth-httplib2==0.1.0
googleapis-common-protos==1.56.4
httplib2==0.20.4
idna==3.3
oauth2client==4.1.3
protobuf==4.21.2
psycopg2-binary==2.9.3
pyaes==1.6.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==3.0.9
requests==2.28.1
rsa==4.8
six==1.16.0
sqlparse==0.4.2
telebot==0.0.4
Telethon==1.24.0
uritemplate==4.1.1
urllib3==1.26.10
